requirements.txt
aiohttp==3.8.4
aioresponses==0.7.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
colorama==0.4.6
coverage==7.2.6
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.10.0
python-dotenv==1.0.0
yarl==1.9.2
